#109: abstracted token management out of my slices page
